+++ app/Http/Controllers/socios/directivoscontroller.php
@@ -19,6 +19,8 @@ class directivoscontroller extends Controller
     public function index()
     {
         //
+        if(!auth()->user()->can('ver directivos'))
+            return response ()->view ('errors.403');
         $directivos = Cargos_directivo::all();
         return view('socios.basicos.directivos')->with('directivos',$directivos);
     }
@@ -43,13 +45,15 @@ class directivoscontroller extends Controller
     {
         //
         if ($request->ajax()) {
+            if(!auth()->user()->can('crear directivos'))
+                return response ()->view ('errors.403-content',[],403);
             $cargo_directivo = Cargos_directivo::create([
                 'cargo_directivo'=>  strtoupper($request->cargo_directivo)
             ]);
             if ($cargo_directivo) {
-                return response()->json(['success' => 'true','message'=>'Se Registro Correctamente']);
+                return response()->json(['success' => true,'message'=>'Se Registro Correctamente']);
             } else {
-                return response()->json(['success' => 'false','message'=>'No se registro ningun datos']);
+                return response()->json(['success' =>false,'message'=>'No se registro ningun datos']);
             }
         }
     }
@@ -88,14 +92,16 @@ class directivoscontroller extends Controller
     public function update(Requests\socios\createCargo_directivorequest $request, $id)
     {
         //
-        $cargo_directivo = Cargos_directivo::FindOrFail($id);
-        $cargo_directivo->fill($request->all());
-        $resul = $cargo_directivo->save();
-        if($resul){
-            return response()->json(['success'=>'true','message'=>'Se actualizaron correctamente los datos']);
-        }
-        else{
-            return response()->json(['success'=>'false','message'=>'No se Actualizaron los datos']);
+        if($request->ajax()) {
+            if(!auth()->user()->can('editar directivos'))
+                return response ()->view ('errors.403-content',[],403);
+            $cargo_directivo = Cargos_directivo::FindOrFail($id);
+            $cargo_directivo->cargo_directivo=  strtoupper($request->cargo_directivo);
+            $resul = $cargo_directivo->save();
+            if ($resul)
+                return response()->json(['success' => true, 'message' => 'Se actualizaron correctamente los datos']);
+            else
+                return response()->json(['success' => false, 'message' => 'No se Actualizaron los datos']);
         }
     }
 
@@ -108,15 +114,13 @@ class directivoscontroller extends Controller
     public function destroy($id)
     {
         //
+        if(auth()->user()->can('eliminar directivos')){
         $cargo_directivo = Cargos_directivo::FindOrFail($id);
         $resul = $cargo_directivo->delete();
-        if($resul)
-        {
-            return response()->json(['success'=>'true']);
-        }
-        else
-        {
-            return response()->json(['success'=>'false']);
+        if($resul) return response()->json(['success'=>true]);
+        
+        else return response()->json(['success'=>false]);
+        
         }
     }
 }
